Extract NewBannerDialog component
